\name{cubeRoot}
\alias{cubeRoot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Cube Root
}
\description{
Cube root function defined as \code{sign(x) * abs(x)^oneThird}.
}
\usage{
cubRoot(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Numeric vector or array
}
}
\value{Numeric vector or array
}
\author{
Long Qu
}
\examples{
cubeRoot(-3)
}
\keyword{ math }

